+++ slitmaskgui/app.py
@@ -29,27 +29,22 @@ logging.basicConfig(
 from slitmaskgui.target_list_widget import TargetDisplayWidget
 from slitmaskgui.mask_gen_widget import MaskGenWidget
 from slitmaskgui.menu_bar import MenuBar
-from slitmaskgui.mask_viewer import interactiveSlitMask, WavelengthView
+from slitmaskgui.mask_widgets.slitmask_view import interactiveSlitMask
+from slitmaskgui.mask_widgets.waveband_view import WavelengthView
 from slitmaskgui.sky_viewer import SkyImageView
 from slitmaskgui.mask_configurations import MaskConfigurationsWidget
 from slitmaskgui.slit_position_table import SlitDisplay
-from slitmaskgui.mask_view_tab_bar import TabBar
-from PyQt6.QtCore import Qt, QSize, pyqtSlot
-from PyQt6.QtGui import QFontDatabase
+from slitmaskgui.mask_widgets.mask_view_tab_bar import TabBar
+from PyQt6.QtCore import Qt, pyqtSlot
 from PyQt6.QtWidgets import (
     QApplication,
     QMainWindow,
     QVBoxLayout,
     QHBoxLayout,
     QWidget,
-    QLabel,
     QSizePolicy,
     QSplitter,
-    QLayout,
-    QTreeWidgetItem,
-    QTreeWidget,
     QTabWidget,
-    QComboBox
 
 
 )
