+++ slitmaskgui/mask_widgets/mask_view_tab_bar.py
